feat: xr physical movement
Previously, XR movement was best-effort and had some restrictions and odd behavior.

This change improves XR so that the player physics capsule moves with you when you physically move in vr, and when snap turning it pivots correctly at the players position. Now, if you physically walk of a high ledge the player will actually fall, and if you try to walk through a wall, physics will push back.